Night 4/7 (checkpoint): gas becomes a mixture, and water evaporates below boiling
A gas cell now holds a mixture (grams per species, plus liquid mist), so
two gases can share a cell and mix. gas::advect and gas::expand are gone:
bulk flow (with momentum) and per-species interdiffusion replace them.
New src/vapour.rs derives each vapour's saturation curve from the boiling
point and latent heat already on its transition (Clausius-Clapeyron), and
runs evaporation, surface condensation/dew, mist and rain on it.

Forced along the way, each by a measured failure:
- Scalar f32 -> f64 (a microgram leaving a gram cell was resolved to ~10%;
  residuals drifted 1e-6..6e-5).
- Condensation relaxes to the equilibrium moved by its own latent heat
  (loop gain 3.6 had the still's head space oscillating above 410 K).
- Liquids touching gas evaporate instead of threshold-boiling (surface
  boiling at the 1-atm point in a pressurised pot pumped heat uphill).
- Gas flow keeps per-face momentum (one relaxation sweep made a 2-cell
  lyne arm need ~2 atm of drive).
- Boussinesq thermal expansion for liquids (pots turn over).
- Bubbles and bursts (small or steep over-pressured pockets) are left
  alone by the vapour rules until flow lets them expand.
- A liquid may always displace gas even if the gas moved this step, and
  same-liquid thermal overturns don't spend cells (terrarium pool slope).
- Phase-change abandon leak (+0.33 J per surfacing cell) fixed.

Scenarios: chamber shows CO2 sinking then mixing (and gets a headless
runner); still gets a relief valve, wider lyne, preheated head space;
terrarium rebuilt as a 365 K spring under a pool with a meadow, reachable
juniper and a gentle fountain. Pages step at the shared SIM_DT (0.05 s)
with a frame budget. Performance: still step ~1.8 ms -> ~0.75 ms native.

Tests: 144 lib (release and debug), 6 e2e green.

+++ src/math.rs
@@ -35,7 +35,23 @@
 /// quantity over a very long run proves `f32` drifts past an acceptable
 /// tolerance), that is a new decision to make there, in the open, against a
 /// measured failure — not a silent reversal of this one.
-pub type Scalar = f32;
+///
+/// **Reversed, 2026-09-10 (night 4), against a measured failure.** Gas cells
+/// became mixtures and liquids began to evaporate below boiling
+/// (`src/vapour.rs`), which means routinely moving a microgram of water out
+/// of a cell holding a gram of it. An `f32` gram has a unit in the last
+/// place of 1.2e-7 g, so each such transfer was resolved to about one part
+/// in ten, and the error went straight into the conservation residuals: a
+/// pond evaporating for 3000 steps drifted its world's mass by 1e-6
+/// relative, and a small warm-pond-under-a-cold-lid water cycle drifted
+/// energy by 6e-5 over 6000 — six times the tolerance the tests already
+/// allowed for rounding. Night 3's journal had seen the leading edge of this
+/// in the terrarium (an energy residual creeping toward 1e-5) and called it
+/// rounding worth watching; it was. The alternative — keeping `f32` and
+/// arranging every transfer so its rounding lands on the smaller cell — fixes
+/// mass but not the temperature solves, and makes every rule harder to read.
+/// The GPU argument above is real but hypothetical; the drift was measured.
+pub type Scalar = f64;
 
 /// A 2D vector of [`Scalar`]s, used for world-space positions, velocities,
 /// forces, and displacements.
